moves the page files to a 'pages' directory
